src/requirements.txt
asgiref==3.5.1
attrs==21.4.0
beautifulsoup4==4.11.1
boto3==1.24.61
botocore==1.27.61
certifi==2021.10.8
cfgv==3.3.1
charset-normalizer==2.0.12
cloudant==2.15.0
coreapi==2.3.3
coreschema==0.0.4
distlib==0.3.4
Django==4.0.4
django-bootstrap4==22.1
django-environ==0.8.1
django-storages==1.13.1
djangorestframework==3.13.1
drf-spectacular==0.23.1
factory-boy==3.2.1
Faker==13.6.0
filelock==3.6.0
flake8==4.0.1
gunicorn==20.1.0
identify==2.5.0
idna==3.3
inflection==0.5.1
iniconfig==1.1.1
itypes==1.2.0
Jinja2==3.1.2
jmespath==1.0.1
jsonschema==4.4.0
MarkupSafe==2.1.1
mccabe==0.6.1
mysqlclient==2.1.1
nodeenv==1.6.0
packaging==23.0
parameterized==0.8.1
platformdirs==2.5.2
pluggy==1.0.0
pre-commit==2.18.1
psycopg2-binary==2.9.3
py==1.11.0
pycodestyle==2.8.0
pyflakes==2.4.0
pyparsing==3.0.8
pyrsistent==0.18.1
pytest==7.1.2
pytest-django==4.5.2
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
requests==2.31
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
s3transfer==0.6.0
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
toml==0.10.2
tomli==2.0.1
uritemplate==4.1.1
urllib3==1.26.9
virtualenv==20.14.1
django-extended-choices==1.3.3